# services/api-gateway/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Microservices API Gateway with command and query services. Supports CQRS pattern with separate command and query endpoints.",
        "title": "KKM API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@kkm-project.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/catalogs-query": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of catalogs",
                "tags": [
                    "Catalogs"
                ],
                "summary": "List catalogs",
                "parameters": [
                    {
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/models.APIResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/models.Catalog"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/catalogs-query/filter": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of catalogs with filter criteria",
                "tags": [
                    "Catalogs"
                ],
                "summary": "List catalogs with filter",
                "parameters": [
                    {
                        "description": "Name filter",
                        "name": "name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Number/code filter",
                        "name": "number",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "TNVED code filter",
                        "name": "tnved_code",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "GKED code filter",
                        "name": "gked_code",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Search text",
                        "name": "search_text",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Sort field (name, number)",
                        "name": "sort_field",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Sort order (ASC/DESC)",
                        "name": "sort_order",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/models.APIResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/models.Catalog"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1"
        },
        {
            "url": "https://localhost:8080/api/v1"
        }
    ],
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "description": "Type \"Bearer\" followed by a space and JWT token.",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "models.APIError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "details": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "models.APIResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "error": {
                        "$ref": "#/components/schemas/models.APIError"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/models.MetaData"
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "models.Catalog": {
                "type": "object",
                "properties": {
                    "category": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "integer"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "unit": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "integer"
                    }
                }
            },
...
            "models.MetaData": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer"
                    },
                    "page_size": {
                        "type": "integer"
                    },
                    "total_count": {
                        "type": "integer"
                    },
                    "total_pages": {
                        "type": "integer"
                    }
                }
            },
...
        }
    }
}